Extract upper-level dependencies from Janet, rename/sort main Janet methods
