+++ janet-base-body/build.gradle
@@ -1,0 +1,8 @@
+apply plugin: 'java'
+
+sourceCompatibility = 1.6
+targetCompatibility = 1.6
+
+dependencies {
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+}
