+++ janet-converters/base/build.gradle
@@ -1,0 +1,9 @@
+apply plugin: 'java'
+
+sourceCompatibility = 1.6
+targetCompatibility = 1.6
+
+dependencies {
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile project(':janet-base-body')
+}
