+++ janet-converters/gson/build.gradle
@@ -1,0 +1,9 @@
+apply plugin: 'java'
+
+sourceCompatibility = 1.6
+targetCompatibility = 1.6
+
+dependencies {
+    compile project(':janet-converters:base')
+    compile 'com.google.code.gson:gson:+'
+}
