+++ janet-http/adapter-compiler/build.gradle
@@ -4,6 +4,7 @@ targetCompatibility = '1.6'
 sourceCompatibility = '1.6'
 
 dependencies {
+    compile project(':janet-base-body')
     compile project(':janet-http:adapter')
     compile 'com.squareup:javapoet:1.1.0'
     compile 'com.google.auto.service:auto-service:1.0-rc2'
