+++ janet-http/adapter/src/main/java/io/techery/janet/http/model/Response.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import io.techery.janet.model.ActionBody;
+import io.techery.janet.body.ActionBody;
 
 public final class Response {
     private final String url;
