+++ janet-http/clients/android/src/main/java/io/techery/janet/android/ApacheClient.java
@@ -27,8 +27,8 @@ import java.util.List;
 import io.techery.janet.http.model.Header;
 import io.techery.janet.http.model.Request;
 import io.techery.janet.http.model.Response;
-import io.techery.janet.model.ActionBody;
-import io.techery.janet.model.BytesArrayBody;
+import io.techery.janet.body.ActionBody;
+import io.techery.janet.body.BytesArrayBody;
 
 
 public class ApacheClient implements io.techery.janet.http.HttpClient {
