+++ janet/src/main/java/io/techery/janet/Interceptor.java
@@ -1,0 +1,7 @@
+package io.techery.janet;
+
+import rx.functions.Action;
+
+public interface Interceptor {
+    void intercept(Action action);
+}
