--- janet/src/main/java/io/techery/janet/JanetException.java
@@ -1,11 +1,0 @@
-package io.techery.janet;
-
-public class JanetException extends RuntimeException {
-
-    public JanetException() {
-    }
-
-    public JanetException(String message) {
-        super(message);
-    }
-}
