+++ janet/src/main/java/io/techery/janet/JanetInternalException.java
@@ -1,0 +1,11 @@
+package io.techery.janet;
+
+public class JanetInternalException extends RuntimeException {
+
+    public JanetInternalException() {
+    }
+
+    public JanetInternalException(String message) {
+        super(message);
+    }
+}
