+++ janet/src/main/java/io/techery/janet/JanetServerException.java
@@ -1,3 +1,3 @@
 package io.techery.janet;
 
-public class JanetServerException extends JanetException {}
+public class JanetServerException extends JanetInternalException {}
