--- janet/src/main/java/io/techery/janet/converter/Converter.java
@@ -1,12 +1,0 @@
-package io.techery.janet.converter;
-
-import java.lang.reflect.Type;
-
-import io.techery.janet.model.ActionBody;
-
-public interface Converter {
-
-    Object fromBody(ActionBody body, Type type);
-
-    ActionBody toBody(Object object);
-}
