--- janet/src/main/java/io/techery/janet/model/BytesArrayBody.java
@@ -1,21 +1,0 @@
-package io.techery.janet.model;
-
-import java.io.IOException;
-
-public class BytesArrayBody extends ActionBody {
-
-    private final byte[] bytes;
-
-    public BytesArrayBody(String mimetype, byte[] bytes) {
-        super(mimetype);
-        if(bytes == null){
-            throw new NullPointerException("bytes == null");
-        }
-        this.bytes = bytes;
-    }
-
-    @Override
-    public byte[] getContent() throws IOException {
-        return bytes;
-    }
-}
