+++ samples/sample-java/src/main/java/io/techery/janet/sample/SimpleService.java
@@ -2,10 +2,10 @@ package io.techery.janet.sample;
 
 import com.google.gson.Gson;
 
-import io.techery.janet.ActionStateSubscriber;
+import io.techery.janet.helper.ActionStateSubscriber;
 import io.techery.janet.HttpActionAdapter;
 import io.techery.janet.Janet;
-import io.techery.janet.JanetPipe;
+import io.techery.janet.ActionPipe;
 import io.techery.janet.gson.GsonConverter;
 import io.techery.janet.okhttp.OkClient;
 import rx.Observable;
@@ -20,20 +20,20 @@ public class SimpleService {
                 .addInterceptor(System.out::println)
                 .build();
 
-        JanetPipe<UsersAction> usersExecutor = janet.createExecutor(UsersAction.class);
-        JanetPipe<UserReposAction> userReposExecutor = janet.createExecutor(UserReposAction.class);
+        ActionPipe<UsersAction> usersPipe = janet.createPipe(UsersAction.class);
+        ActionPipe<UserReposAction> userReposPipe = janet.createPipe(UserReposAction.class);
 
-        usersExecutor.observeActions()
+        usersPipe.observeActions()
                 .filter(BaseAction::isSuccess)
                 .subscribe(
                         action -> System.out.println("received " + action),
                         System.err::println
                 );
 
-        usersExecutor.createObservable(new UsersAction())
+        usersPipe.createObservable(new UsersAction())
                 .filter(state -> state.action.isSuccess())
                 .flatMap(state -> Observable.<User>from(state.action.response).first())
-                .flatMap(user -> userReposExecutor.createObservable(new UserReposAction(user.getLogin())))
+                .flatMap(user -> userReposPipe.createObservable(new UserReposAction(user.getLogin())))
                 .subscribe(new ActionStateSubscriber<UserReposAction>()
                         .onSuccess(action -> System.out.println("repos request finished " + action))
                         .onFail(throwable -> System.err.println("repos request throwable " + throwable))
